check is transaction exist befere delete

+++ server/orpc/routers/transaction.ts
@@ -1,3 +1,4 @@
+import { ORPCError } from '@orpc/server';
 import z from 'zod';
 import { publicProcedure } from '@/server/orpc';
 import { prisma } from '@/server/prisma';
@@ -21,7 +22,7 @@ export const transactionRouter = {
     .route({ method: 'POST', path: '/transactions' })
     .input(transactionSchema.omit({ id: true, createAt: true }))
     .output(transactionSchema)
-    .handler(async ({ input }) => {
+    .handler(({ input }) => {
       return prisma.transaction.create({
         data: {
           text: input.text,
@@ -35,6 +36,16 @@ export const transactionRouter = {
     .input(z.object({ id: z.number() }))
     .output(transactionSchema)
     .handler(async ({ input }) => {
+      const transaction = await prisma.transaction.findUnique({
+        where: { id: input.id },
+      });
+
+      if (!transaction) {
+        throw new ORPCError('NOT_FOUND', {
+          message: `Transaction with ID ${input.id} not found`,
+        });
+      }
+
       return prisma.transaction.delete({
         where: { id: input.id },
       });
